extract execution of kwmc commands to its own function
